refactor: LottoNumbers 삭제 및 LottoTicket 클래스가 기능 위임
