+++ src/test/java/lotto/domain/LottoTicketTest.java
@@ -7,27 +7,17 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class LottoNumbersTest {
-    LottoNumbers lottoNumbers, winningNumbers;
+public class LottoTicketTest {
+    LottoTicket lottoTicket;
+    LottoTicket winningNumbers;
 
     @BeforeEach
     void init() {
-        List<Integer> numberList = Arrays.asList(1, 2, 3, 4, 5, 6);
-        List<LottoNumber> lottoNumberList = new ArrayList<>();
-        numberList.forEach(number -> lottoNumberList.add(new LottoNumber(number)));
-        lottoNumbers = new LottoNumbers(lottoNumberList);
-
-        List<Integer> winningNumberList = Arrays.asList(6, 5, 4, 3, 2, 1);
-        List<LottoNumber> winningLottoNumberList = new ArrayList<>();
-        winningNumberList.forEach(number -> winningLottoNumberList.add(new LottoNumber(number)));
-        winningNumbers = new LottoNumbers(winningLottoNumberList);
+        lottoTicket = new LottoTicket("1, 2, 3, 4, 5, 6");
+        winningNumbers = new LottoTicket("6, 5, 4, 3, 2, 1");
     }
 
     @ParameterizedTest
@@ -35,7 +25,7 @@ public class LottoNumbersTest {
     @DisplayName("당첨 로또 번호 null 이거나 비어있을 경우 Exception 발생 확인")
     void validateNullOrEmpty(String input) {
         assertThatThrownBy(() -> {
-            new LottoNumbers(input);
+            new LottoTicket(input);
         }).isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -44,7 +34,7 @@ public class LottoNumbersTest {
     @DisplayName("당첨 로또 번호 중에 중복되는 번호가 존재할 경우 Exception 발생 확인")
     void validateUnique(String input) {
         assertThatThrownBy(() -> {
-            new LottoNumbers(input);
+            new LottoTicket(input);
         }).isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -53,7 +43,7 @@ public class LottoNumbersTest {
     @DisplayName("당첨 로또 번호가 6개가 아닐 경우 Exception 발생 확인")
     void validateSize(String input) {
         assertThatThrownBy(() -> {
-            new LottoNumbers(input);
+            new LottoTicket(input);
         }).isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -62,19 +52,19 @@ public class LottoNumbersTest {
     @DisplayName("당첨 로또 번호 중에 범위 내(1-45) 에 있지 않은 번호가 존재할 경우 Exception 발생 확인")
     void validateRange(String input) {
         assertThatThrownBy(() -> {
-            new LottoNumbers(input);
+            new LottoTicket(input);
         }).isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
     @DisplayName("구매한 로또 번호와 당첨 번호의 매칭 개수 확인")
     void matchCount() {
-        assertThat(lottoNumbers.matchCount(winningNumbers)).isEqualTo(6);
+        assertThat(lottoTicket.matchCount(winningNumbers)).isEqualTo(6);
     }
 
     @DisplayName("당첨 순위 확인")
     @Test
     void rank() {
-        assertThat(lottoNumbers.rank(winningNumbers)).isEqualTo(LottoRank.FIRST);
+        assertThat(lottoTicket.rank(winningNumbers)).isEqualTo(LottoRank.FIRST);
     }
 }
